+++ src/rxjs.js
@@ -1,8 +1,10 @@
 import * as t from 'babel-types'
-import {makeProxy, subjectIdentifier} from './rx'
+import {makeProxy, getSubjectLiteral} from './rx'
 
-export const addImports = (node, scope) => {
-  if (!scope.hasBinding('Subject')){
+export const addImports = (node, scope, options) => {
+  const subjectLiteral = getSubjectLiteral(options)
+  const subjectIdentifier = t.identifier(subjectLiteral)
+  if (!scope.hasBinding(subjectLiteral)){
     const subjectImportDeclaration = t.importDeclaration([
       t.importSpecifier(subjectIdentifier, subjectIdentifier)
     ], t.stringLiteral('rxjs'));