+++ test/fixtures/rx-subject-imported/transformed.js
@@ -1,3 +1,4 @@
+import { ReplaySubject } from 'rx';
 import { Component1 } from './component1';
 import { Component2 } from './component2';
 import { Subject } from 'rx';